rename keeper package to store
