+++ store/interface.go
@@ -1,4 +1,4 @@
-package keeper //TODO: to rename to store
+package store
 
 // Store describes the API of keeper of Services.
 type Store interface {
